WT-18427 Extract the scrub-image wait into a shared eviction helper
Both scrub-eviction test files polled the scrub-image gauges to zero after
a drop with their own copy of the loop, differing only in deadline and
sleep. Move scrub_images() and the wait onto eviction_util, which both
files now inherit, and standardise on the more generous deadline so slow
variants have margin (the measured drain is ~0.1s).

Drop the local get_stat() in the 02 file: WiredTigerTestCase already
provides an equivalent, so it was a redundant reimplementation.

Also remove two more copies of the dead underflow assertions and their
incorrect rationale. As with the copy corrected earlier, the accounting
clamps a negative decrement to zero before anything can read the gauge, so
these could never fire, and both branches at each site already assert the
stronger exact value.
