+++ test/suite/test_checkpoint_scrub_evict02.py
@@ -37,14 +37,13 @@
 # underflow, and the cache reports non-zero counters when it is destroyed.
 
 import threading
-import time
 
-import wttest
+from eviction_util import eviction_util
 from wiredtiger import stat, WiredTigerError, wiredtiger_strerror, WT_ROLLBACK
 from wtscenario import make_scenarios
 from wtthread import Thread
 
-class test_checkpoint_scrub_evict02(wttest.WiredTigerTestCase):
+class test_checkpoint_scrub_evict02(eviction_util):
     uri = 'table:scrub_evict02'
     nrows = 5000
 
@@ -63,16 +62,6 @@ class test_checkpoint_scrub_evict02(wttest.WiredTigerTestCase):
                 'checkpoint_threads=%d,'
                 'eviction_dirty_target=80,eviction_dirty_trigger=90,'
                 'eviction=(checkpoint_scrub_eviction=on)' % self.ckpt_threads)
-
-    def get_stat(self, statistic, uri=None):
-        cursor = self.session.open_cursor('statistics:' if uri is None else 'statistics:' + uri)
-        value = cursor[statistic][2]
-        cursor.close()
-        return value
-
-    def scrub_images(self):
-        return (self.get_stat(stat.conn.cache_scrub_image_pages),
-          self.get_stat(stat.conn.cache_scrub_image_bytes))
 
     def create(self, config=''):
         self.session.create(self.uri, 'key_format=i,value_format=S' + config)
@@ -139,10 +128,8 @@ class test_checkpoint_scrub_evict02(wttest.WiredTigerTestCase):
     def test_images_released_when_page_discarded(self):
         """Dropping the table frees pages that still hold a retained image.
 
-        A clean tree's handle is marked dead by the drop and its pages are freed later by the
-        sweep server, not synchronously as part of the drop call, so this polls rather than
-        checking right away. Speed up the sweep scan so the poll resolves quickly instead of
-        waiting on the default interval."""
+        Speed up the sweep scan so the wait resolves quickly instead of waiting on the default
+        interval."""
         self.settle()
         self.write('y' * 100)
         self.session.checkpoint()
@@ -151,11 +138,7 @@ class test_checkpoint_scrub_evict02(wttest.WiredTigerTestCase):
         self.conn.reconfigure('file_manager=(close_scan_interval=1)')
         self.session.drop(self.uri)
 
-        deadline = time.time() + 10
-        while self.scrub_images() != (0, 0):
-            self.assertLess(time.time(), deadline,
-                'scrub images were not released within the deadline')
-            time.sleep(0.5)
+        self.wait_for_scrub_images_released()
 
     def test_images_released_by_later_reconciliation(self):
         """A page dirtied and reconciled again releases the image its last checkpoint retained.
